backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-storage>=2.16.0
google-auth>=2.29.0
pydantic>=2.7.0
python-multipart>=0.0.9
websockets>=12.0
